Add files via upload
Employ management system using C++ is a menu-driven program that allows us to add, update, delete and search records of an employee working in an organization. The program employee management system stores employee ID, name, post, department, and salary of the employee. Initially, it has no data.
here we can update the records whenever necessary and also search for the updated records of the employees.